Unify RX and TX queue management and isolation
This is Work In Progress, but feedback is welcome since it
is a pretty significant change.

This patch improves existing queue management framework such
that both tx and rx queue management and placement is done
in a unified way.

It also introduces a subdivision of the thread execution from
the perspective of packet reception and transmission. This new
feature allows isolation of packets received from different
queues by ensuring that packets that should not be mixed
(e.g. coming from different RX queues in order to preserve
RSS hashing) are processed seperatly. Similar mechanism is
added on transmission.

This work was done in order to solve a situation where it
was important that a VM receives packets through multiple
virtio queues (In order to balance the load between cores).
For instance, a 8 cores VMs would need to receive packets
from 8 different queues. Since VPP does not perform RSS
hashing, the solution was to enable RSS on the physical
interface (with 8 different DPDK queues). But since there
was no way to isolate the packet processing from the multiple
queues, the only working solution was to use 8 VPP worker
threads. Which is a waste of cores since a single, or a
couple, of VPP workers are able to saturate all the VM queues.

This patch solves this issue by allowing each thread to
independently process packets from different input queues,
and send them to different tx queues. In the previous example,
the thread execution is virtually cut into 8 rss slots, with
each slot being assigned a single RX queue, and a single TX
queue.

More information can be found in src/vnet/devices/device_rss.md

Change-Id: Ic6b2342fd0515d04832a933dae84b370ab3c0e8f
Signed-off-by: Pierre Pfister <ppfister@cisco.com>